+++ src/main/java/nysleep/model/Accommodation.java
@@ -4,21 +4,21 @@ package nysleep.model;
 import java.util.List;
 
 public class Accommodation {
-    private long id;
+    private int id;
     private String name;
     private String neighborhood;
     private List<String> imagesURL;
     private int numBeds;
-    private float rating;
+    private double rating;
     private int numReviews;
     private String propertyType;
     private int numRooms;
     private List<String> amenities;
-    private float price;
-    private long renterId;
+    private double price;
+    private int renterId;
     private List<Reservation> reservations;
 
-    public Accommodation(long id, String name, String neighborhood, List<String> imagesURL, int numBeds, float rating, int numReviews, String propertyType, int numRooms, List<String> amenities, float price, long renterId, List<Reservation> reservations) {
+    public Accommodation(int id, String name, String neighborhood, List<String> imagesURL, int numBeds, double rating, int numReviews, String propertyType, int numRooms, List<String> amenities, double price, int renterId, List<Reservation> reservations) {
         this.id = id;
         this.name = name;
         this.neighborhood = neighborhood;
@@ -50,9 +50,9 @@ public class Accommodation {
 
     public void setNumBeds(int numBeds) {this.numBeds = numBeds;}
 
-    public float getRating() {return rating;}
+    public double getRating() {return rating;}
 
-    public void setRating(float rating) {this.rating = rating;}
+    public void setRating(double rating) {this.rating = rating;}
 
     public int getNumReviews() {return numReviews;}
 
@@ -70,13 +70,13 @@ public class Accommodation {
 
     public void setAmenities(List<String> amenities) {this.amenities = amenities;}
 
-    public float getPrice() {return price;}
+    public double getPrice() {return price;}
 
-    public void setPrice(float price) {this.price = price;}
+    public void setPrice(double price) {this.price = price;}
 
-    public long getRenterId() {return renterId;}
+    public int getRenterId() {return renterId;}
 
-    public void setRenterId(long renterId) {this.renterId = renterId;}
+    public void setRenterId(int renterId) {this.renterId = renterId;}
 
     public List<Reservation> getReservations() {
         return reservations;
@@ -86,11 +86,11 @@ public class Accommodation {
         this.reservations = reservations;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
